pyproject.toml
[build-system]
requires = ["setuptools >= 80.9.0"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-ai"
version = "0.0.1"
authors = [
  { name="Tiago Matos", email="tiagolofi@example.com" },
]
description = ""
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
  "python-dotenv", "pandas", "openai", "openpyxl"
]
license = "Apache-2.0"
license-files = ["LICEN[CS]E*"]

[project.urls]
Homepage = "https://github.com/tiagolofi/prompt-ai"
Issues = "https://github.com/tiagolofi/prompt-ai/issues"
    
